# gh-96349: Fix minor performance regression initializing threading.Event