--- a/Misc/NEWS.d/next/Library/2022-08-27-21-26-52.gh-issue-96349.XyYLlO.rst
+++ b/Misc/NEWS.d/next/Library/2022-08-27-21-26-52.gh-issue-96349.XyYLlO.rst
@@ -1,0 +1,1 @@
+Fixed a minor performance regression in :func:`threading.Event.__init__`