<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
gh-NNNNN: Summary of the changes made
```

Where: gh-NNNNN refers to the GitHub issue number.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `main`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `main`.

-->
There was a minor performance regression making it slightly slower to initialize a `threading.Event` on the main branch than it was in 3.10.  I think the slowdown may have been related to zero-cost exceptions.

This PR removes the slowdown (and makes it faster than it was before).

Here's a simple benchmark with pyperf:

3.10:
```
../cpython3.10/python -m pyperf timeit -s 'from threading import Event' 'e = Event()'
.....................
Mean +- std dev: 1.52 us +- 0.02 us
```
Main branch:
```
../cpython/python -m pyperf timeit -s 'from threading import Event' 'e = Event()'
.....................
Mean +- std dev: 1.59 us +- 0.02 us
```
With this PR:
```
../cpython/python -m pyperf timeit -s 'from threading import Event' 'e = Event()'
.....................
Mean +- std dev: 462 ns +- 10 ns
```


<!-- gh-issue-number: gh-96349 -->
* Issue: gh-96349
<!-- /gh-issue-number -->